Rename IOFormat to Ui and add error printers
IOFormat is Ui after all ;)
