+++ text-ui-test/EXPECTED.TXT
@@ -1,20 +1,20 @@
 💬 Hello, Duke here! What can I do for you?
 ─────
-💬 I have gotten you started with an empty task list. Welcome!
+🟡 I have gotten you started with an empty task list. Welcome!
 ─────
-💬 Sorry, I didn't understand what you said :(
+🟡 Sorry, I didn't understand what you said :(
 ─────
-💬 Sorry, I didn't catch that?
+🟡 Sorry, I didn't catch that?
 ─────
 💬 Here are the tasks in your list:
 ─────
-💬 Please give your todo a name.
+🟡 Please give your todo a name.
 ─────
 💬 Got it. I've added this task:
    [T][ ]  borrow book
    Now you have 1 task in the list.
 ─────
-💬 Deadlines are added like this: deadline return book /by Sunday
+🟡 Deadlines are added like this: deadline return book /by Sunday
 ─────
 💬 Got it. I've added this task:
    [D][ ]  return book (by: 30 Aug 2022, 12:34:56)
@@ -24,9 +24,9 @@
    [D][ ]  The significance of this date is left as an exercise for the reader (by: 2 Jan 2006, 15:04:05)
    Now you have 3 tasks in the list.
 ─────
-💬 nussu doesn't look like a date and time to me...
+🟡 nussu doesn't look like a date and time to me...
 ─────
-💬 Events are added like this: event project meeting /at Mon 2-4pm
+🟡 Events are added like this: event project meeting /at Mon 2-4pm
 ─────
 💬 Got it. I've added this task:
    [E][ ]  project meeting (at: Mon 2-4pm)
@@ -38,7 +38,7 @@
    3. [D][ ]  The significance of this date is left as an exercise for the reader (by: 2 Jan 2006, 15:04:05)
    4. [E][ ]  project meeting (at: Mon 2-4pm)
 ─────
-💬 Please give me a task number to mark!
+🟡 Please give me a task number to mark!
 ─────
 💬 Nice! I've marked this task as done:
    [T][X]  borrow book
@@ -52,7 +52,7 @@
    3. [D][ ]  The significance of this date is left as an exercise for the reader (by: 2 Jan 2006, 15:04:05)
    4. [E][X]  project meeting (at: Mon 2-4pm)
 ─────
-💬 Please give me a task number to unmark!
+🟡 Please give me a task number to unmark!
 ─────
 💬 Alright, I've marked this task as not done yet:
    [T][ ]  borrow book
@@ -63,13 +63,13 @@
    3. [D][ ]  The significance of this date is left as an exercise for the reader (by: 2 Jan 2006, 15:04:05)
    4. [E][X]  project meeting (at: Mon 2-4pm)
 ─────
-💬 Hmm... a task with that number doesn't seem to exist. You can see a list of all tasks by saying "list".
+🟡 Hmm... a task with that number doesn't seem to exist. You can see a list of all tasks by saying "list".
 ─────
-💬 Hmm... a task with that number doesn't seem to exist. You can see a list of all tasks by saying "list".
+🟡 Hmm... a task with that number doesn't seem to exist. You can see a list of all tasks by saying "list".
 ─────
-💬 Hmm... a task with that number doesn't seem to exist. You can see a list of all tasks by saying "list".
+🟡 Hmm... a task with that number doesn't seem to exist. You can see a list of all tasks by saying "list".
 ─────
-💬 Hmm... a task with that number doesn't seem to exist. You can see a list of all tasks by saying "list".
+🟡 Hmm... a task with that number doesn't seem to exist. You can see a list of all tasks by saying "list".
 ─────
 💬 Here are the tasks in your list:
    1. [T][ ]  borrow book
@@ -77,13 +77,13 @@
    3. [D][ ]  The significance of this date is left as an exercise for the reader (by: 2 Jan 2006, 15:04:05)
    4. [E][X]  project meeting (at: Mon 2-4pm)
 ─────
-💬 Please give me a task number to delete!
+🟡 Please give me a task number to delete!
 ─────
 💬 OK, I've deleted this task:
    [D][ ]  return book (by: 30 Aug 2022, 12:34:56)
    Now you have 3 tasks in the list.
 ─────
-💬 Hmm... a task with that number doesn't seem to exist. You can see a list of all tasks by saying "list".
+🟡 Hmm... a task with that number doesn't seem to exist. You can see a list of all tasks by saying "list".
 ─────
 💬 Here are the tasks in your list:
    1. [T][ ]  borrow book
